permissionamento simples

+++ routes/web.php
@@ -50,10 +50,10 @@ Route::post('/categoria/create', [CategoriaController::class, 'store'])->name('c
 
 Route::get('/categoria/{id}', [CategoriaController::class, 'show'])->name('categoria.show');
 
-Route::get('/categoria/{id}/edit', [CategoriaController::class, 'edit'])->name('categoria.edit');
-Route::put('/categoria/{id}', [CategoriaController::class, 'update'])->name('categoria.update');
+Route::get('/categoria/{id}/edit', [CategoriaController::class, 'edit'])->name('categoria.edit')->can('is_admin');
+Route::put('/categoria/{id}', [CategoriaController::class, 'update'])->name('categoria.update')->can('is_admin');
 
-Route::delete('/categoria/{id}', [CategoriaController::class, 'destroy'])->name('categoria.destroy');
+Route::delete('/categoria/{id}', [CategoriaController::class, 'destroy'])->name('categoria.destroy')->can('is_admin');
 
 
 // ----------------------------------------- CATEGORIA --------------------------------------
